requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
attrs==25.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.2.0
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.6.0
geoip2==2.9.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
icalendar==5.0.11
idna==3.6
joblib==1.3.2
maxminddb==2.7.0
mcp==1.8.1
mcp-server-time==0.6.2
multidict==6.4.3
numpy==2.2.5
packaging==25.0
propcache==0.3.1
psycopg==3.1.16
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
sse-starlette==2.3.5
starlette==0.46.2
threadpoolctl==3.2.0
twilio==8.10.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.1.0
uvicorn==0.34.2
whitenoise==6.5.0
yarl==1.20.0
xhtml2pdf==0.2.16
weasyprint>=66.0
